Move trash/untrash views to browser directory and ungrok them.
Completely remove grok from opengever.trash.
