+++ opengever/trash/configure.zcml
@@ -3,7 +3,6 @@
     xmlns:plone="http://namespaces.plone.org/plone"
     xmlns:i18n="http://namespaces.zope.org/i18n"
     xmlns:monkey="http://namespaces.plone.org/monkey"
-    xmlns:grok="http://namespaces.zope.org/grok"
     xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
     i18n_domain="opengever.trash">
 
@@ -11,6 +10,7 @@
   <i18n:registerTranslations directory="locales" />
 
   <include file="profiles.zcml" />
+  <include file="permissions.zcml" />
 
   <include package="plone.behavior" file="meta.zcml" />
 
@@ -21,11 +21,6 @@
   <include package=".upgrades" />
 
   <include package=".browser" />
-
-  <!-- Include custom permissions -->
-  <include file="permissions.zcml" />
-
-  <grok:grok package="." />
 
   <plone:behavior
       title="ITrashable Behavior"
